updated to irrlicht 1.8.3
updated Irrlicht source to the latest release

+++ src/dep/src/irrlicht/C3DSMeshFileLoader.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2002-2011 Nikolaus Gebhardt
+// Copyright (C) 2002-2012 Nikolaus Gebhardt
 // This file is part of the "Irrlicht Engine".
 // For conditions of distribution and use, see copyright notice in irrlicht.h
 
@@ -109,7 +109,7 @@ private:
 			MaterialName = o.MaterialName;
 			faceCount = o.faceCount;
 			faces = new u16[faceCount];
-			for (u32 i=0; i<faceCount; ++i)
+			for (u16 i=0; i<faceCount; ++i)
 				faces[i] = o.faces[i];
 		}
